+++ opensearchpy/helpers/__init__.py
@@ -25,8 +25,12 @@
 #  under the License.
 
 
-import sys
-
+from .._async.helpers.actions import (
+    async_bulk,
+    async_reindex,
+    async_scan,
+    async_streaming_bulk,
+)
 from .actions import (
     _chunk_actions,
     _process_bulk_chunk,
@@ -56,16 +60,8 @@ __all__ = [
     "AWSV4SignerAsyncAuth",
     "RequestsAWSV4SignerAuth",
     "Urllib3AWSV4SignerAuth",
+    "async_scan",
+    "async_bulk",
+    "async_reindex",
+    "async_streaming_bulk",
 ]
-
-
-# Asyncio only supported on Python 3.6+
-if sys.version_info >= (3, 6):
-    from .._async.helpers.actions import (
-        async_bulk,
-        async_reindex,
-        async_scan,
-        async_streaming_bulk,
-    )
-
-    __all__ += ["async_scan", "async_bulk", "async_reindex", "async_streaming_bulk"]